fix: revert CommunityId to u16, use manual storage version bump
Changing CommunityId from u16 to u32 breaks all existing on-chain
storage encoding. Instead, skip the frame-contrib referenda-tracks
migration and just set storage version to 1 directly since the
existing data layout is compatible.

Also:
- Remove now-unnecessary SKIP_PALLET_REVIVE_FIXTURES env vars
- Add wasm32-unknown-unknown target to check CI for contracts fixtures

+++ common/src/lib.rs
@@ -13,7 +13,7 @@ pub use payment_id::PaymentId;
 #[cfg(feature = "runtime")]
 pub use multilocation_asset_id::runtime::AsFungibleAssetLocation;
 
-pub type CommunityId = u32;
+pub type CommunityId = u16;
 pub type MembershipId = u32;
 
 pub mod listings {
